+++ changelog.md
@@ -13,7 +13,7 @@ _Not yet scheduled_
   * unapproved-user-experimental.yaml renamed from "unauthz"
 
 ## Infrastructure
-* none
+* AgencyPublicDataExport - see [/doc/data-export](/doc/data-export).
 
 
 ---